feat: Add a already registered arns name to deployment
